Function radix_engine_stores::hash_tree::put_at_next_version

pub fn put_at_next_version<S: TreeStore>(
    node_tier_store: &S,
    node_root_version: Option<Version>,
    database_updates: &DatabaseUpdates
) -> Hash

Inserts a new set of nodes at version node_root_version + 1 into the “3-Tier JMT” persisted within the given TreeStore. In a traditional JMT, this inserts a new leaf node for each given “change”, together with an entire new “parent chain” leading from that leaf to a new root node (common for all of them). In our instantiation of the JMT, we first update all touched Substate-Tier JMTs, then we update all touched Partition-Tier JMTs and then we update the single ReNode-Tier tree. All nodes that became stale precisely due to this (i.e. not any previous) operation will be reported before the function returns (see WriteableTreeStore::record_stale_node). Returns the hash of the newly-created root (i.e. representing state at version node_root_version + 1).

Panics

Panics if a root node for node_root_version does not exist. The caller should use None to denote an empty, initial state of the tree (i.e. inserting at version 1).