Module radix_engine_stores::hash_tree::jellyfish

Structs

JellyfishMerkleTree

The Jellyfish Merkle tree data structure. See crate for description.

NibbleRangeIterator 🔒

An iterator that iterates the index range (inclusive) of each different nibble at given nibble_idx of all the keys in a sorted key-value pairs which have the identical Hash prefix (up to nibble_idx).

StaleNodeIndex

Indicates a node becomes stale since stale_since_version.

TreeUpdateBatch

Constants

SANITY_NIBBLE_LIMIT 🔒