Struct radix_engine_stores::hash_tree::jellyfish::JellyfishMerkleTree

pub struct JellyfishMerkleTree<'a, R, P> {
    reader: &'a R,
    phantom_value: PhantomData<P>,
}

The Jellyfish Merkle tree data structure. See crate for description.

Fields
reader: &'a Rphantom_value: PhantomData<P>
Implementations

impl<'a, R: 'a + TreeReader<P>, P: Clone> JellyfishMerkleTree<'a, R, P>

pub fn new(reader: &'a R) -> Self

Creates a JellyfishMerkleTree backed by the given TreeReader.

fn get_hash( node_key: &NodeKey, node: &Node<P>, hash_cache: &Option<&HashMap<NibblePath, Hash>> ) -> Hash

Get the node hash from the cache if cache is provided, otherwise (for test only) compute it.

pub fn batch_put_value_set( &self, value_set: Vec<(&LeafKey, Option<&(Hash, P)>)>, node_hashes: Option<&HashMap<NibblePath, Hash>>, persisted_version: Option<Version>, version: Version ) -> Result<(Hash, TreeUpdateBatch<P>), StorageError>

For each value set: Returns the new nodes and values in a batch after applying value_set. For example, if after transaction T_i the committed state of tree in the persistent storage looks like the following structure:

             S_i
            /   \
           .     .
          .       .
         /         \
        o           x
       / \
      A   B
       storage (disk)

where A and B denote the states of two adjacent accounts, and x is a sibling subtree of the path from root to A and B in the tree. Then a value_set produced by the next transaction T_{i+1} modifies other accounts C and D exist in the subtree under x, a new partial tree will be constructed in memory and the structure will be:

                S_i      |      S_{i+1}
               /   \     |     /       \
              .     .    |    .         .
             .       .   |   .           .
            /         \  |  /             \
           /           x | /               x'
          o<-------------+-               / \
         / \             |               C   D
        A   B            |
          storage (disk) |    cache (memory)

With this design, we are able to query the global state in persistent storage and generate the proposed tree delta based on a specific root hash and value_set. For example, if we want to execute another transaction T_{i+1}', we can use the tree S_i in storage and apply the value_set of transaction T_{i+1}. Then if the storage commits the returned batch, the state S_{i+1} is ready to be read from the tree by calling get_with_proof. Anything inside the batch is not reachable from public interfaces before being committed.

fn batch_insert_at( &self, node_key: &NodeKey, version: Version, kvs: &[(&LeafKey, Option<&(Hash, P)>)], depth: usize, hash_cache: &Option<&HashMap<NibblePath, Hash>>, batch: &mut TreeUpdateBatch<P> ) -> Result<Option<Node<P>>, StorageError>

fn insert_at_child( &self, node_key: &NodeKey, internal_node: &InternalNode, version: Version, kvs: &[(&LeafKey, Option<&(Hash, P)>)], left: usize, right: usize, depth: usize, hash_cache: &Option<&HashMap<NibblePath, Hash>>, batch: &mut TreeUpdateBatch<P> ) -> Result<(Nibble, Option<Node<P>>), StorageError>

fn batch_update_subtree_with_existing_leaf( &self, node_key: &NodeKey, version: Version, existing_leaf_node: LeafNode<P>, kvs: &[(&LeafKey, Option<&(Hash, P)>)], depth: usize, hash_cache: &Option<&HashMap<NibblePath, Hash>>, batch: &mut TreeUpdateBatch<P> ) -> Result<Option<Node<P>>, StorageError>

fn batch_update_subtree( &self, node_key: &NodeKey, version: Version, kvs: &[(&LeafKey, Option<&(Hash, P)>)], depth: usize, hash_cache: &Option<&HashMap<NibblePath, Hash>>, batch: &mut TreeUpdateBatch<P> ) -> Result<Option<Node<P>>, StorageError>

pub fn get_with_proof( &self, key: &LeafKey, version: Version ) -> Result<(Option<(Hash, P, Version)>, SparseMerkleProof), StorageError>

Returns the value (if applicable) and the corresponding merkle proof.

pub fn get_with_proof_ext( &self, key: &LeafKey, version: Version ) -> Result<(Option<(Hash, P, Version)>, SparseMerkleProofExt), StorageError>

pub fn get_range_proof( &self, rightmost_key_to_prove: &LeafKey, version: Version ) -> Result<SparseMerkleRangeProof, StorageError>

Gets the proof that shows a list of keys up to rightmost_key_to_prove exist at version.

fn get_root_node(&self, version: Version) -> Result<Node<P>, StorageError>

pub fn get_root_hash(&self, version: Version) -> Result<Hash, StorageError>

pub fn get_leaf_count(&self, version: Version) -> Result<usize, StorageError>

pub fn get_all_nodes_referenced( &self, key: NodeKey ) -> Result<Vec<NodeKey>, StorageError>

fn get_all_nodes_referenced_impl( &self, key: NodeKey, out_keys: &mut Vec<NodeKey> ) -> Result<(), StorageError>