Struct radix_engine_stores::hash_tree::types::LeafNode

pub struct LeafNode<P> {
    leaf_key: LeafKey,
    value_hash: Hash,
    payload: P,
    version: Version,
}

Leaf node, capturing the value hash and carrying an arbitrary payload.

Fields
leaf_key: LeafKeyvalue_hash: Hashpayload: Pversion: Version
Implementations

impl LeafNode<Version>

pub fn from(key: &NodeKey, leaf_node: &TreeLeafNode) -> Self

impl<P: Clone> LeafNode<P>

pub fn new( leaf_key: LeafKey, value_hash: Hash, payload: P, version: Version ) -> Self

Creates a new leaf node.

pub fn leaf_key(&self) -> &LeafKey

Gets the key.

pub fn value_hash(&self) -> Hash

Gets the associated value hash.

pub fn payload(&self) -> &P

Gets the payload.

pub fn version(&self) -> Version

Gets the version.

pub fn leaf_hash(&self) -> Hash

Gets the leaf’s hash (not to be confused with a value_hash()). This hash incorporates the node’s key and the value’s hash, in order to capture certain changes within a sparse merkle tree (consider 2 trees, both containing a single element with the same value, but stored under different keys - we want their root hashes to differ).