Struct radix_engine_stores::hash_tree::types::NibblePath

pub struct NibblePath {
    num_nibbles: usize,
    bytes: Vec<u8>,
}

NibblePath defines a path in Merkle tree in the unit of nibble (4 bits).

Fields
num_nibbles: usize

Indicates the total number of nibbles in bytes. Either bytes.len() * 2 - 1 or bytes.len() * 2.

bytes: Vec<u8>

The underlying bytes that stores the path, 2 nibbles per byte. If the number of nibbles is odd, the second half of the last byte must be 0.

Implementations

impl NibblePath

pub fn new_even(bytes: Vec<u8>) -> Self

Creates a new NibblePath from a vector of bytes assuming each byte has 2 nibbles.

pub fn new_odd(bytes: Vec<u8>) -> Self

Similar to new() but asserts that the bytes have one less nibble.

pub fn push(&mut self, nibble: Nibble)

Adds a nibble to the end of the nibble path.

pub fn pop(&mut self) -> Option<Nibble>

Pops a nibble from the end of the nibble path.

pub fn last(&self) -> Option<Nibble>

Returns the last nibble.

fn get_bit(&self, i: usize) -> bool

Get the i-th bit.

pub fn get_nibble(&self, i: usize) -> Nibble

Get the i-th nibble.

pub fn bits(&self) -> BitIterator<'_>

Get a bit iterator iterates over the whole nibble path.

pub fn nibbles(&self) -> NibbleIterator<'_>

Get a nibble iterator iterates over the whole nibble path.

pub fn num_nibbles(&self) -> usize

Get the total number of nibbles stored.

pub fn is_empty(&self) -> bool

Returns true if the nibbles contains no elements.

pub fn bytes(&self) -> &[u8]

Get the underlying bytes storing nibbles.

pub fn truncate(&mut self, len: usize)